+++ NEWS.md
@@ -93,6 +93,8 @@
 
 * Add Carlsson Petri (2021) liraglutide PK model
 * Add a variety of indirect response models
+* Add a variety of tumor growth inhibition models and move all oncology models
+  into a new model database directory
 * `cp` and related `cpddSd` and `cppropSd` were renamed to `Cc`, `CcAddSd` and
   `CcPropSd` (fix #70).